+++ man-roxygen/loo-and-psis-references.R
@@ -1,0 +1,8 @@
+#' @references Vehtari, A., Gelman, A., and Gabry, J. (2016a). Practical
+#'   Bayesian model evaluation using leave-one-out cross-validation and WAIC.
+#'   \emph{Statistics and Computing}. Advance online publication.
+#'   doi:10.1007/s11222-016-9696-4. arXiv preprint:
+#'   \url{http://arxiv.org/abs/1507.04544/}
+#'
+#' @references Vehtari, A., Gelman, A., and Gabry, J. (2016b). Pareto smoothed
+#'   importance sampling. arXiv preprint: \url{http://arxiv.org/abs/1507.02646/}