+++ tests/testthat/test_compare.R
@@ -1,13 +1,15 @@
 library(loo)
 options(loo.cores = 1)
 
+SW <- suppressWarnings
+
 context("compare")
 test_that("compare throws appropriate errors", {
   set.seed(123)
-  x1 <- waic(matrix(rnorm(5000), 100, 50))
-  x2 <- waic(matrix(rnorm(4000), 100, 40))
-  x3 <- waic(matrix(rnorm(3000), 100, 30))
-  x4 <- waic(matrix(rnorm(5000), 100, 50))
+  x1 <- SW(waic(matrix(rnorm(5000), 100, 50)))
+  x2 <- SW(waic(matrix(rnorm(4000), 100, 40)))
+  x3 <- SW(waic(matrix(rnorm(3000), 100, 30)))
+  x4 <- SW(waic(matrix(rnorm(5000), 100, 50)))
 
   expect_error(loo::compare(x1, list(1,2,3)), "class 'loo'")
   expect_error(loo::compare(x1),
@@ -23,20 +25,23 @@ test_that("compare throws appropriate errors", {
 })
 test_that("compare returns expected result (2 models)", {
   set.seed(123)
-  x1 <- loo(matrix(rnorm(5000), 100, 50))
-  x2 <- loo(matrix(rnorm(5000), 100, 50))
+  x1 <- SW(loo(matrix(rnorm(5000), 100, 50)))
+  x2 <- SW(loo(matrix(rnorm(5000), 100, 50)))
   diff_val <- compare(x1, x2)
   diff_ans <- structure(c(-0.49758869857943333059, 1.22288821000758018975),
                         .Names = c("elpd_diff", "se"),
                         class = "compare.loo")
   expect_equal(diff_val, diff_ans)
+
+  # specifying objects via ... and via arg x gives equal results
+  expect_equivalent(compare(x1, x2), compare(x = list(x1, x2)))
 })
 
 test_that("compare returns expected result (3 models)", {
   set.seed(123)
-  x1 <- loo(matrix(rnorm(5000), 100, 50))
-  x2 <- loo(matrix(rnorm(5000), 100, 50))
-  x3 <- loo(matrix(rnorm(5000), 100, 50))
+  x1 <- SW(loo(matrix(rnorm(5000), 100, 50)))
+  x2 <- SW(loo(matrix(rnorm(5000), 100, 50)))
+  x3 <- SW(loo(matrix(rnorm(5000), 100, 50)))
   diff_val <- compare(x1, x2, x3)
 
   nms <- c("looic", "se_looic", "elpd_loo", "se_elpd_loo", "p_loo", "se_p_loo")
@@ -51,4 +56,7 @@ test_that("compare returns expected result (3 models)", {
               .Dimnames = list(c("x1", "x3", "x2"), nms),
               class = c("compare.loo", "matrix"))
   expect_equal(diff_val, diff_ans)
+
+  # specifying objects via ... and via arg x gives equal results
+  expect_equivalent(compare(x1, x2, x3), compare(x = list(x1, x2, x3)))
 })