+++ tests/testthat/test_psislw.R
@@ -18,8 +18,20 @@ test_that("psislw returns expected results", {
 
 test_that("psislw handles special cases, throws appropriate errors/warnings", {
   expect_warning(psis <- psislw(x[, 1], wcp = 0.01),
-                 regexp = "Too few tail samples")
+                 regexp = "All tail values are the same. Weights are truncated but not smoothed")
+  expect_warning(psislw(x[, 1], wcp = 0.01),
+                 regexp = "Some Pareto k diagnostic values are too high. See help('pareto-k-diagnostic') for details",
+                 fixed = TRUE)
   expect_true(is.infinite(psis$pareto_k))
   expect_error(psislw(wcp = 0.2),
                regexp = "'lw' or 'llfun' and 'llargs' must be specified")
+})
+
+test_that("psislw_warnings helper works properly", {
+  k <- c(0, 0.1, 0.55, 0.75)
+  expect_silent(psislw_warnings(k[1:2]))
+  expect_warning(psislw_warnings(k[1:3]),
+                 "Some Pareto k diagnostic values are slightly high")
+  expect_warning(psislw_warnings(k),
+                 "Some Pareto k diagnostic values are too high")
 })